[ADD] split avocado tests into multiple files.
